New: Automatic import of Half-Life 1 and 2 mods #769, #788
* Clean up usage of Steam App IDs, and implement Game ID. Steam's Game ID is
a 64-bit representation of an app, usually only used locally. The lower 32-
bits are just the App ID, but the rest can be used to identify a mod, or a
non-Steam shortcut. In the case of a mod, the App ID portion is still used
for Steam to know which app to launch. In most places, a Game ID can be
used in place of an App ID since they have the same value for "normal" apps.

This shouldn't have any net change, but will allow implementation of Source
mod support in the Playnite Steam plugin.

* Fix usages of App IDs to be uint instead of int, to match Steam.
* Implement Steam Game ID for local usages (not online).
* Switch from steam://run to rungame, to support running by Game ID.
* Switch Playnite Game ID to use Steam Game ID instead of App ID for more
clear identification of app.

* Add support for Steam Source mods.

This isn't perfect, but works fairly well. Source mods are scanned for in a
similar method to how the Steam client does it, with details also pulled in
the same way (from gameinfo.txt).

* More updates to Steam mod support, including HL1 mods.

* Abstract mod logic to ModInfo class.
* Add support for HL1 mods.
* Point metadata to ModInfo for mods.
* Add error dialog for trying to install or uninstall a mod.

* Add support for legacy url_* liblist fields for HL1 mods.

* Cleanup, round 1.

* Add support for icons for Steam mods.

* Refactor SteamLibrary to be a member of SteamMetadataProvider.

* Fix mod icon lookup issues.

* Fix HL1 mod blacklist check being inverted.