Import LastPlayed during Steam import
`LastActivity` is now imported during `GetLibraryGames(...). However, if
a public library is being imported but
settings.IdSource == SteamIdSource.Name, we won't have an API key, so we
can't convert the `AccountName` to an `AccountId`. In that case
importing LastActivity data is skipped it.

Errors importing LastActivity during import are silently logged and
ignored. Errors are still shown when maunally importing LastActivity for
Steam.