Fix games not getting tagged as uninstalled
With the "Import installed games" setting the GameDatabase.ImportGames
method only updated properties for games that are installed on the
machine. Games that were uninstalled were still left tagged as
installed, which could lead to errors when trying to launch a
non-existent game.

This patch adds a check to make sure that games are now marked as
uninstalled if they are no longer installed in the client. It also does
a minor code cleanup.