Modify Loc Strings related to playtime sync
I propose modifying three strings to be more clear-

1. ´Force sync play time´ -> ´Force sync Play Time for supported libraries´ (Makes it clear that not every library is supported)
2. ´Always download recorded play time from remote library during library sync, ignoring time recorded by Playnite.´ -> ´Always download recorded play time from remote library during library sync, ignoring time recorded by Playnite. GOG, Origin and Steam libraries are supported when the user is authenticated.´ (To specify exactly what libraries are supported and to explain that the user must be currently authenticated in each respective library for Playtime Sync to work.
3. ´Playtime is tracked and stored in seconds.´ -> ´Play Time is tracked and stored in seconds.´ (For consistency since the other strings have a space between ´Play´ and ´Time´. I also capitalized the first letter of each word.

Just as another suggeston, I was thinking that the terms ´Play Time´ and ´Time Played´ in the different strings could be unified using one of the two, since they have the same meaning. If this was the case, I'm more inclined to use ´Time Played"

I propose this because I've seen confusion on which libraries can sync playtime and some playtime sync problems that have been reported were because the user was not authenticated, as seen in #1481 and on discord different times. It should also be mentioned that the GOG sync is currently not working as mentioned in #1474