Add info for switching active theme (#3193)
This info for setting your active theme should really/also be on the "How to: Themes" page but I couldn't figure out how to edit that.
https://github.com/JosefNemec/Playnite/wiki/How-to:-Themes

Info on how to change your active/applied theme isn't available anywhere that I could find. I could only find a reddit post with someone else asking the same question. We both assumed you could toggle a theme active in the Addons>Installed>Themes window (like how Firefox etc does it). [That would also be a nice addition to the program.]
My next guess was that in the main dropdown menu there was a Theme submenu, or View>Theme or Add-ons>Theme or Extensions>Theme or something where I could select my active theme. I was wrong again so I had to look through help docs and then just do an internet search.

It is also not intuitive that you have to activate Fullscreen mode separately before you can choose the theme for that view in settings (it would be helpful to show 2 options in the regular settings, one for Desktop Theme and one for Fullscreen Theme. I do see that there's a completely different Settings UI in fullscreen mode though so it's probably fine to leave that side as is). [As a sidenote, I also assumed there would be a fullscreen button on the top-right like the steam big picture mode button in the steam ui.]

Anyway, thanks for your work on this great program!